item and collection crud done

+++ src/main/java/com/itransition/coursework/tag/TagView.java
@@ -1,7 +1,9 @@
 package com.itransition.coursework.tag;
 
+import lombok.Data;
+
 public interface TagView {
-    Integer getId();
+    Long getId();
 
     String getName();
 }
